+++ crates/wdl-lint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Fixed
+
+* Updated shellcheck logic that erroneously flagged placeholders that are quoted ([#541](https://github.com/stjude-rust-labs/wdl/pull/541)).
+
 #### Removed
 
 * Removed the `OutputSection` lint rule ([#532](https://github.com/stjude-rust-labs/wdl/pull/532)).